requirements.txt
google-generativeai
requests
twitchio
gtts
python-dotenv